ks拓扑成功安装 测试用户态当中

+++ audio/sysvad/TabletAudioSample/AudioVolumeControl/AudioEQTypes.h
@@ -1,50 +1,42 @@
 #pragma once
+//
+// AudioEQTypes.h —— EQ 基础类型
+// 只放数据结构，不放 IOCTL/分发声明。类型来自 <ntddk.h>，不要自定义 BYTE 等。
+// 
 
-// 最大 EQ 段数定义
-#define EQ_BANDS 12
-#define MAX_EQ_BANDS EQ_BANDS
+#include <ntddk.h>
+
+#define EQ_BANDS      12
+#define MAX_EQ_BANDS  EQ_BANDS
 
 #pragma pack(push, 1)
 
-// =========================================================================
-// EQ 系数结构体（b0, b1, b2, a1, a2）
-// =========================================================================
-typedef struct _EQBandCoeffs
-{
-    int b0, b1, b2;
-    int a1, a2;
+// 单段双二阶（biquad）系数（Q15）
+typedef struct _EQBandCoeffs {
+    LONG b0, b1, b2;   // 前向系数
+    LONG a1, a2;       // 反馈系数（a0 假定为 1.0）
 } EQBandCoeffs;
 
-// =========================================================================
-// EQ 状态结构体（x1, x2, y1, y2）：用于每个声道保存滤波状态
-// =========================================================================
-typedef struct _EQBandRuntime
-{
-    int x1, x2;
-    int y1, y2;
+// 每段每声道的历史状态
+typedef struct _EQBandRuntime {
+    LONG x1, x2;       // 输入延迟
+    LONG y1, y2;       // 输出延迟
 } EQBandRuntime;
 
-// =========================================================================
-// EQ 参数结构体：用于 IOCTL 接收一整组系数（可选）
-// =========================================================================
-typedef struct _EQCoeffParams
-{
-    int BandCount;
-    EQBandCoeffs Bands[EQ_BANDS];
+// 一次性下发/读取的一组（最多 12 段）系数
+typedef struct _EQCoeffParams {
+    LONG          BandCount;          // 1..EQ_BANDS
+    EQBandCoeffs  Bands[EQ_BANDS];    // 前 BandCount 个有效
 } EQCoeffParams;
 
 #pragma pack(pop)
 
-// =========================================================================
-// 外部变量声明（在 .cpp 中定义）
-// =========================================================================
+// 仅供 EQ 模块实现使用（在 .cpp 中定义）
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-extern EQBandCoeffs   g_EqBands[EQ_BANDS];          // 每段 EQ 的系数
-extern EQBandRuntime  g_EqStates[EQ_BANDS][2];      // 每段 EQ 的左右声道状态
-
+extern EQBandCoeffs  g_EqBands[EQ_BANDS];
+extern EQBandRuntime g_EqStates[EQ_BANDS][2];  // [band][0=L,1=R]
 #ifdef __cplusplus
 }
 #endif
